fix: Rename folders
