+++ server/_passportConfig/guestV2Strategy.ts
